Format changes
Unfortunately, I hadn't noticed these issued before.  Most have to do with unnecessarily obscuring or distorting the template graphics.  It is okay to cover them if there is no reasonable alternative, but usually, things can be formatted to avoid the background.  It is not acceptable to distort the logos (this seems to happen to the IDEAS logo a lot).