Add presentations from prior tutorials
These tutorials pre-date the use of git to manage the presentations and the creation of this repo.  They include both the ppt and pdf files, since we don't have them in the history of the ppts in the main directory.

For the present, I'm not capturing tutorial presentations for which I don't have ppt files.